Move teardown logic into system
