+++ radix-engine/src/kernel/kernel_callback_api.rs
@@ -7,10 +7,12 @@ use crate::kernel::substate_io::SubstateDevice;
 use crate::track::interface::{IOAccess, NodeSubstates};
 use crate::track::{BootStore, StoreCommitInfo, Track};
 use radix_engine_interface::api::field_api::LockFlags;
+use radix_engine_interface::blueprints::transaction_processor::InstructionOutput;
 use radix_substate_store_interface::db_key_mapper::SpreadPrefixKeyMapper;
 use radix_substate_store_interface::interface::SubstateDatabase;
 use radix_transactions::model::Executable;
 use radix_transactions::prelude::PreAllocatedAddress;
+use crate::transaction::{CostingParameters, TransactionFeeDetails, TransactionFeeSummary, TransactionResult};
 
 pub trait CallFrameReferences {
     fn root() -> Self;
@@ -163,6 +165,18 @@ pub trait KernelCallbackObject: Sized {
         Y: KernelApi<Self>;
 
     fn on_teardown2(&mut self, store_commit_info: StoreCommitInfo) -> Result<(), RuntimeError>;
+
+    fn on_teardown3<S: SubstateDatabase>(
+        self,
+        track: Track<S, SpreadPrefixKeyMapper>,
+        executable: &Executable,
+        result: Result<Vec<InstructionOutput>, TransactionExecutionError>,
+    ) -> (
+        CostingParameters,
+        TransactionFeeSummary,
+        Option<TransactionFeeDetails>,
+        TransactionResult,
+    );
 
     fn on_pin_node(&mut self, node_id: &NodeId) -> Result<(), RuntimeError>;
