+++ radix-engine/src/system/system_callback.rs
@@ -22,7 +22,7 @@ use crate::system::module::{InitSystemModule, SystemModule};
 use crate::system::system::SystemService;
 use crate::system::system_callback_api::SystemCallbackObject;
 use crate::system::system_modules::auth::AuthModule;
-use crate::system::system_modules::costing::{CostingModule, FeeTable, FinalizationCostingEntry, StorageType, SystemLoanFeeReserve};
+use crate::system::system_modules::costing::{CostingModule, FeeReserveFinalizationSummary, FeeTable, FinalizationCostingEntry, FinalizingFeeReserve, StorageType, SystemLoanFeeReserve};
 use crate::system::system_modules::execution_trace::ExecutionTraceModule;
 use crate::system::system_modules::kernel_trace::KernelTraceModule;
 use crate::system::system_modules::limits::LimitsModule;
@@ -30,8 +30,8 @@ use crate::system::system_modules::transaction_runtime::TransactionRuntimeModule
 use crate::system::system_modules::{EnabledModules, SystemModuleMixer};
 use crate::system::system_substates::KeyValueEntrySubstate;
 use crate::system::system_type_checker::{BlueprintTypeTarget, KVStoreTypeTarget};
-use crate::track::{BootStore, CommitableSubstateStore, StoreCommitInfo, Track};
-use crate::transaction::{CostingParameters, LimitParameters, SystemOverrides};
+use crate::track::{BootStore, CommitableSubstateStore, StoreCommitInfo, to_state_updates, Track, TrackFinalizeError};
+use crate::transaction::{AbortResult, CommitResult, CostingParameters, FeeDestination, FeeSource, LimitParameters, reconcile_resource_state_and_events, RejectResult, StateUpdateSummary, SystemOverrides, SystemStructure, TransactionFeeDetails, TransactionFeeSummary, TransactionOutcome, TransactionResult, TransactionResultType};
 use radix_blueprint_schema_init::RefTypes;
 use radix_engine_interface::api::field_api::LockFlags;
 use radix_engine_interface::api::ClientObjectApi;
@@ -45,13 +45,14 @@ use radix_engine_interface::blueprints::hooks::OnVirtualizeInput;
 use radix_engine_interface::blueprints::hooks::OnVirtualizeOutput;
 use radix_engine_interface::blueprints::identity::IDENTITY_BLUEPRINT;
 use radix_engine_interface::blueprints::package::*;
-use radix_engine_interface::blueprints::transaction_processor::{
-    TRANSACTION_PROCESSOR_BLUEPRINT, TRANSACTION_PROCESSOR_RUN_IDENT,
-};
+use radix_engine_interface::blueprints::transaction_processor::{InstructionOutput, TRANSACTION_PROCESSOR_BLUEPRINT, TRANSACTION_PROCESSOR_RUN_IDENT};
 use radix_substate_store_interface::{db_key_mapper::SpreadPrefixKeyMapper, interface::*};
-use radix_transactions::model::{Executable, PreAllocatedAddress};
-use crate::blueprints::consensus_manager::{ConsensusManagerField, ConsensusManagerStateFieldPayload};
+use radix_transactions::model::{Executable, PreAllocatedAddress, TransactionIntentHash};
+use crate::blueprints::consensus_manager::{ConsensusManagerField, ConsensusManagerStateFieldPayload, ConsensusManagerValidatorRewardsFieldPayload};
+use crate::blueprints::resource::{BurnFungibleResourceEvent, FungibleVaultBalanceFieldPayload, FungibleVaultBalanceFieldSubstate, FungibleVaultField};
+use crate::blueprints::resource::fungible_vault::{DepositEvent, PayFeeEvent};
 use crate::blueprints::transaction_tracker::{TransactionStatus, TransactionStatusV1, TransactionTrackerSubstate};
+use crate::system::system_db_reader::SystemDatabaseReader;
 
 pub const BOOT_LOADER_SYSTEM_SUBSTATE_FIELD_KEY: FieldKey = 1u8;
 
@@ -225,6 +226,375 @@ impl<C: SystemCallbackObject> System<C> {
 
         Ok(())
     }
+
+    fn determine_result_type(
+        interpretation_result: Result<Vec<InstructionOutput>, TransactionExecutionError>,
+        fee_reserve: &mut SystemLoanFeeReserve,
+    ) -> TransactionResultType {
+        // A `SuccessButFeeLoanNotRepaid` error is issued if a transaction finishes before
+        // the SYSTEM_LOAN_AMOUNT is reached (which trigger a repay event) and even though
+        // enough fee has been locked.
+        //
+        // Do another `repay` try during finalization to remedy it.
+        let final_repay_result = fee_reserve.repay_all();
+
+        match interpretation_result {
+            Ok(output) => match final_repay_result {
+                Ok(_) => TransactionResultType::Commit(Ok(output)), // success and system loan repaid fully
+                Err(e) => {
+                    if let Some(abort_reason) = e.abortion() {
+                        TransactionResultType::Abort(abort_reason.clone())
+                    } else {
+                        TransactionResultType::Reject(RejectionReason::SuccessButFeeLoanNotRepaid)
+                    }
+                }
+            },
+            Err(e) => match e {
+                TransactionExecutionError::BootloadingError(e) => {
+                    TransactionResultType::Reject(RejectionReason::BootloadingError(e))
+                }
+                TransactionExecutionError::RuntimeError(e) => {
+                    if let Some(abort_reason) = e.abortion() {
+                        TransactionResultType::Abort(abort_reason.clone())
+                    } else {
+                        if fee_reserve.fully_repaid() {
+                            TransactionResultType::Commit(Err(e))
+                        } else {
+                            TransactionResultType::Reject(
+                                RejectionReason::ErrorBeforeLoanAndDeferredCostsRepaid(e),
+                            )
+                        }
+                    }
+                }
+            },
+        }
+    }
+
+
+    fn finalize_fees<S: SubstateDatabase>(
+        track: &mut Track<S, SpreadPrefixKeyMapper>,
+        fee_reserve: SystemLoanFeeReserve,
+        is_success: bool,
+        free_credit: Decimal,
+    ) -> (
+        FeeReserveFinalizationSummary,
+        IndexMap<NodeId, Decimal>,
+        Vec<(EventTypeIdentifier, Vec<u8>)>,
+    ) {
+        let mut events = Vec::<(EventTypeIdentifier, Vec<u8>)>::new();
+
+        // Distribute royalty
+        for (recipient, amount) in fee_reserve.royalty_cost_breakdown().clone() {
+            let node_id = recipient.vault_id();
+            let substate_key = FungibleVaultField::Balance.into();
+            let mut vault_balance = track
+                .read_substate(&node_id, MAIN_BASE_PARTITION, &substate_key)
+                .unwrap()
+                .as_typed::<FungibleVaultBalanceFieldSubstate>()
+                .unwrap()
+                .into_payload()
+                .into_latest();
+            vault_balance.put(LiquidFungibleResource::new(amount));
+            let updated_substate_content =
+                FungibleVaultBalanceFieldPayload::from_content_source(vault_balance)
+                    .into_unlocked_substate();
+            track
+                .set_substate(
+                    node_id,
+                    MAIN_BASE_PARTITION,
+                    substate_key,
+                    IndexedScryptoValue::from_typed(&updated_substate_content),
+                    &mut |_| -> Result<(), ()> { Ok(()) },
+                )
+                .unwrap();
+            events.push((
+                EventTypeIdentifier(
+                    Emitter::Method(node_id, ModuleId::Main),
+                    DepositEvent::EVENT_NAME.to_string(),
+                ),
+                scrypto_encode(&DepositEvent { amount }).unwrap(),
+            ));
+        }
+
+        // Take fee payments
+        let fee_reserve_finalization = fee_reserve.finalize();
+        let mut fee_payments: IndexMap<NodeId, Decimal> = index_map_new();
+        let mut required = fee_reserve_finalization.total_cost();
+        let mut collected_fees = LiquidFungibleResource::new(Decimal::ZERO);
+        for (vault_id, mut locked, contingent) in
+        fee_reserve_finalization.locked_fees.iter().cloned().rev()
+        {
+            let amount = if contingent {
+                if is_success {
+                    Decimal::min(locked.amount(), required)
+                } else {
+                    Decimal::zero()
+                }
+            } else {
+                Decimal::min(locked.amount(), required)
+            };
+
+            // NOTE: Decimal arithmetic operation safe unwrap.
+            // No chance to overflow considering current costing parameters
+
+            // Take fees
+            collected_fees.put(locked.take_by_amount(amount).unwrap());
+            required = required.checked_sub(amount).unwrap();
+
+            // Refund overpayment
+            let mut vault_balance = track
+                .read_substate(
+                    &vault_id,
+                    MAIN_BASE_PARTITION,
+                    &FungibleVaultField::Balance.into(),
+                )
+                .unwrap()
+                .as_typed::<FungibleVaultBalanceFieldSubstate>()
+                .unwrap()
+                .into_payload()
+                .into_latest();
+            vault_balance.put(locked);
+            let updated_substate_content =
+                FungibleVaultBalanceFieldPayload::from_content_source(vault_balance)
+                    .into_unlocked_substate();
+            track
+                .set_substate(
+                    vault_id,
+                    MAIN_BASE_PARTITION,
+                    FungibleVaultField::Balance.into(),
+                    IndexedScryptoValue::from_typed(&updated_substate_content),
+                    &mut |_| -> Result<(), ()> { Ok(()) },
+                )
+                .unwrap();
+
+            // Record final payments
+            let entry = fee_payments.entry(vault_id).or_default();
+            *entry = entry.checked_add(amount).unwrap();
+
+            events.push((
+                EventTypeIdentifier(
+                    Emitter::Method(vault_id, ModuleId::Main),
+                    PayFeeEvent::EVENT_NAME.to_string(),
+                ),
+                scrypto_encode(&PayFeeEvent { amount }).unwrap(),
+            ));
+        }
+        // Free credit is locked first and thus used last
+        if free_credit.is_positive() {
+            let amount = Decimal::min(free_credit, required);
+            collected_fees.put(LiquidFungibleResource::new(amount));
+            required = required.checked_sub(amount).unwrap();
+        }
+
+        let to_proposer = fee_reserve_finalization.to_proposer_amount();
+        let to_validator_set = fee_reserve_finalization.to_validator_set_amount();
+        let to_burn = fee_reserve_finalization.to_burn_amount();
+
+        // Sanity checks
+        assert!(
+            fee_reserve_finalization.total_bad_debt_in_xrd == Decimal::ZERO,
+            "Bad debt is non-zero: {}",
+            fee_reserve_finalization.total_bad_debt_in_xrd
+        );
+        assert!(
+            required == Decimal::ZERO,
+            "Locked fee does not cover transaction cost: {} required",
+            required
+        );
+        let remaining_collected_fees = collected_fees.amount().checked_sub(fee_reserve_finalization.total_royalty_cost_in_xrd /* royalty already distributed */).unwrap();
+        let to_distribute = to_proposer
+            .checked_add(to_validator_set)
+            .unwrap()
+            .checked_add(to_burn)
+            .unwrap();
+        assert!(
+            remaining_collected_fees  == to_distribute,
+            "Remaining collected fee isn't equal to amount to distribute (proposer/validator set/burn): {} != {}",
+            remaining_collected_fees,
+            to_distribute,
+        );
+
+        if !to_proposer.is_zero() || !to_validator_set.is_zero() {
+            // Fetch current leader
+            // TODO: maybe we should move current leader into validator rewards?
+            let substate: FieldSubstate<ConsensusManagerStateFieldPayload> = track
+                .read_substate(
+                    CONSENSUS_MANAGER.as_node_id(),
+                    MAIN_BASE_PARTITION,
+                    &ConsensusManagerField::State.into(),
+                )
+                .unwrap()
+                .as_typed()
+                .unwrap();
+            let current_leader = substate.into_payload().into_latest().current_leader;
+
+            // Update validator rewards
+            let substate: FieldSubstate<ConsensusManagerValidatorRewardsFieldPayload> = track
+                .read_substate(
+                    CONSENSUS_MANAGER.as_node_id(),
+                    MAIN_BASE_PARTITION,
+                    &ConsensusManagerField::ValidatorRewards.into(),
+                )
+                .unwrap()
+                .as_typed()
+                .unwrap();
+
+            let mut rewards = substate.into_payload().into_latest();
+
+            if let Some(current_leader) = current_leader {
+                let entry = rewards.proposer_rewards.entry(current_leader).or_default();
+                *entry = entry.checked_add(to_proposer).unwrap()
+            } else {
+                // If there is no current leader, the rewards go to the pool
+            };
+            let vault_node_id = rewards.rewards_vault.0 .0;
+
+            track
+                .set_substate(
+                    CONSENSUS_MANAGER.into_node_id(),
+                    MAIN_BASE_PARTITION,
+                    ConsensusManagerField::ValidatorRewards.into(),
+                    IndexedScryptoValue::from_typed(&FieldSubstate::new_unlocked_field(
+                        ConsensusManagerValidatorRewardsFieldPayload::from_content_source(rewards),
+                    )),
+                    &mut |_| -> Result<(), ()> { Ok(()) },
+                )
+                .unwrap();
+
+            // Put validator rewards into the vault
+            let total_amount = to_proposer.checked_add(to_validator_set).unwrap();
+            let mut vault_balance = track
+                .read_substate(
+                    &vault_node_id,
+                    MAIN_BASE_PARTITION,
+                    &FungibleVaultField::Balance.into(),
+                )
+                .unwrap()
+                .as_typed::<FungibleVaultBalanceFieldSubstate>()
+                .unwrap()
+                .into_payload()
+                .into_latest();
+            vault_balance.put(collected_fees.take_by_amount(total_amount).unwrap());
+            let updated_substate_content =
+                FungibleVaultBalanceFieldPayload::from_content_source(vault_balance)
+                    .into_unlocked_substate();
+            track
+                .set_substate(
+                    vault_node_id,
+                    MAIN_BASE_PARTITION,
+                    FungibleVaultField::Balance.into(),
+                    IndexedScryptoValue::from_typed(&updated_substate_content),
+                    &mut |_| -> Result<(), ()> { Ok(()) },
+                )
+                .unwrap();
+
+            events.push((
+                EventTypeIdentifier(
+                    Emitter::Method(vault_node_id, ModuleId::Main),
+                    DepositEvent::EVENT_NAME.to_string(),
+                ),
+                scrypto_encode(&DepositEvent {
+                    amount: total_amount,
+                })
+                    .unwrap(),
+            ));
+        }
+
+        if to_burn.is_positive() {
+            events.push((
+                EventTypeIdentifier(
+                    Emitter::Method(XRD.into_node_id(), ModuleId::Main),
+                    "BurnFungibleResourceEvent".to_string(),
+                ),
+                scrypto_encode(&BurnFungibleResourceEvent { amount: to_burn }).unwrap(),
+            ));
+        }
+
+        (fee_reserve_finalization, fee_payments, events)
+    }
+
+    fn update_transaction_tracker<S: SubstateDatabase>(
+        track: &mut Track<S, SpreadPrefixKeyMapper>,
+        next_epoch: Epoch,
+        intent_hash: &TransactionIntentHash,
+        is_success: bool,
+    ) {
+        // Read the intent hash store
+        let transaction_tracker = track
+            .read_substate(
+                TRANSACTION_TRACKER.as_node_id(),
+                MAIN_BASE_PARTITION,
+                &TransactionTrackerField::TransactionTracker.into(),
+            )
+            .unwrap()
+            .as_typed::<FieldSubstate<TransactionTrackerSubstate>>()
+            .unwrap()
+            .into_payload();
+
+        let mut transaction_tracker = transaction_tracker.into_v1();
+
+        // Update the status of the intent hash
+        if let TransactionIntentHash::ToCheck {
+            expiry_epoch,
+            intent_hash,
+        } = intent_hash
+        {
+            if let Some(partition_number) =
+                transaction_tracker.partition_for_expiry_epoch(*expiry_epoch)
+            {
+                track
+                    .set_substate(
+                        TRANSACTION_TRACKER.into_node_id(),
+                        PartitionNumber(partition_number),
+                        SubstateKey::Map(scrypto_encode(intent_hash).unwrap()),
+                        IndexedScryptoValue::from_typed(&KeyValueEntrySubstate::V1(
+                            KeyValueEntrySubstateV1 {
+                                value: Some(if is_success {
+                                    TransactionStatus::V1(TransactionStatusV1::CommittedSuccess)
+                                } else {
+                                    TransactionStatus::V1(TransactionStatusV1::CommittedFailure)
+                                }),
+                                // TODO: maybe make it immutable, but how does this affect partition deletion?
+                                lock_status: LockStatus::Unlocked,
+                            },
+                        )),
+                        &mut |_| -> Result<(), ()> { Ok(()) },
+                    )
+                    .unwrap();
+            } else {
+                panic!("No partition for an expiry epoch")
+            }
+        }
+
+        // Check if all intent hashes in the first epoch have expired, based on the `next_epoch`.
+        //
+        // In this particular implementation, because the transaction tracker coverage is greater than
+        // the max epoch range in transaction header, we must check epoch range first to
+        // ensure we don't store intent hash too far into the future.
+        //
+        // Also, we need to make sure epoch doesn't jump by a large distance.
+        if next_epoch.number()
+            >= transaction_tracker.start_epoch + transaction_tracker.epochs_per_partition
+        {
+            let discarded_partition = transaction_tracker.advance();
+            track.delete_partition(
+                TRANSACTION_TRACKER.as_node_id(),
+                PartitionNumber(discarded_partition),
+            );
+        }
+        track
+            .set_substate(
+                TRANSACTION_TRACKER.into_node_id(),
+                MAIN_BASE_PARTITION,
+                TransactionTrackerField::TransactionTracker.into(),
+                IndexedScryptoValue::from_typed(&FieldSubstate::new_unlocked_field(
+                    TransactionTrackerSubstate::V1(transaction_tracker),
+                )),
+                &mut |_| -> Result<(), ()> { Ok(()) },
+            )
+            .unwrap();
+    }
+
 }
 
 impl<C: SystemCallbackObject> KernelCallbackObject for System<C> {
@@ -488,6 +858,176 @@ impl<C: SystemCallbackObject> KernelCallbackObject for System<C> {
 
         Ok(())
     }
+
+    fn on_teardown3<S: SubstateDatabase>(
+        self,
+        mut track: Track<S, SpreadPrefixKeyMapper>,
+        executable: &Executable,
+        interpretation_result: Result<Vec<InstructionOutput>, TransactionExecutionError>,
+    )  -> (
+        CostingParameters,
+        TransactionFeeSummary,
+        Option<TransactionFeeDetails>,
+        TransactionResult,
+    ) {
+        #[cfg(not(feature = "alloc"))]
+        if self.modules.enabled_modules.contains(EnabledModules::KERNEL_TRACE) {
+            println!("{:-^120}", "Interpretation Results");
+            println!("{:?}", interpretation_result);
+        }
+
+        let execution_trace_enabled = self.modules.enabled_modules.contains(EnabledModules::EXECUTION_TRACE);
+
+        let (mut costing_module, runtime_module, execution_trace_module) = self.modules.unpack();
+
+        let costing_parameters = costing_module.fee_reserve.costing_parameters();
+
+        let fee_details = if let Some(cost_breakdown) = &costing_module.cost_breakdown {
+            let cost_breakdown = cost_breakdown.clone();
+            let execution_cost_breakdown = cost_breakdown
+                .execution_cost_breakdown
+                .into_iter()
+                .map(|(k, v)| (k.to_string(), v))
+                .collect();
+            let finalization_cost_breakdown = cost_breakdown
+                .finalization_cost_breakdown
+                .into_iter()
+                .map(|(k, v)| (k.to_string(), v))
+                .collect();
+            Some(TransactionFeeDetails {
+                execution_cost_breakdown,
+                finalization_cost_breakdown,
+            })
+        } else {
+            None
+        };
+
+        let result_type = Self::determine_result_type(
+            interpretation_result,
+            &mut costing_module.fee_reserve,
+        );
+        match result_type {
+            TransactionResultType::Commit(outcome) => {
+                let is_success = outcome.is_ok();
+
+                // Commit/revert
+                if !is_success {
+                    costing_module.fee_reserve.revert_royalty();
+                    track.revert_non_force_write_changes();
+                }
+
+                // Distribute fees
+                let (fee_reserve_finalization, paying_vaults, finalization_events) =
+                    Self::finalize_fees(
+                        &mut track,
+                        costing_module.fee_reserve,
+                        is_success,
+                        executable.costing_parameters().free_credit_in_xrd,
+                    );
+                let fee_destination = FeeDestination {
+                    to_proposer: fee_reserve_finalization.to_proposer_amount(),
+                    to_validator_set: fee_reserve_finalization.to_validator_set_amount(),
+                    to_burn: fee_reserve_finalization.to_burn_amount(),
+                    to_royalty_recipients: fee_reserve_finalization
+                        .royalty_cost_breakdown
+                        .clone(),
+                };
+
+                // Update intent hash status
+                if let Some(next_epoch) = Self::read_epoch(&mut track) {
+                    Self::update_transaction_tracker(
+                        &mut track,
+                        next_epoch,
+                        executable.intent_hash(),
+                        is_success,
+                    );
+                }
+
+                // Finalize events and logs
+                let (mut application_events, application_logs) =
+                    runtime_module.finalize(is_success);
+                application_events.extend(finalization_events);
+
+                // Finalize execution trace
+                let execution_trace =
+                    execution_trace_module.finalize(&paying_vaults, is_success);
+
+                // Finalize track
+                let (tracked_substates, substate_db) = {
+                    match track.finalize() {
+                        Ok(result) => result,
+                        Err(TrackFinalizeError::TransientSubstateOwnsNode) => {
+                            panic!("System invariants should prevent transient substate from owning nodes");
+                        }
+                    }
+                };
+
+                // Generate state updates from tracked substates
+                // Note that this process will prune invalid reads
+                let (new_node_ids, state_updates) =
+                    to_state_updates::<SpreadPrefixKeyMapper>(tracked_substates);
+
+                // Summarizes state updates
+                let system_structure = SystemStructure::resolve(
+                    substate_db,
+                    &state_updates,
+                    &application_events,
+                );
+                let state_update_summary =
+                    StateUpdateSummary::new(substate_db, new_node_ids, &state_updates);
+
+                // Resource reconciliation does not currently work in preview mode
+                if executable.costing_parameters().free_credit_in_xrd.is_zero() {
+                    let system_reader = SystemDatabaseReader::new_with_overlay(
+                        substate_db,
+                        &state_updates,
+                    );
+                    reconcile_resource_state_and_events(
+                        &state_update_summary,
+                        &application_events,
+                        system_reader,
+                    );
+                }
+
+                (
+                    costing_parameters,
+                    fee_reserve_finalization.into(),
+                    fee_details,
+                    TransactionResult::Commit(CommitResult {
+                        state_updates,
+                        state_update_summary,
+                        fee_source: FeeSource { paying_vaults },
+                        fee_destination,
+                        outcome: match outcome {
+                            Ok(o) => TransactionOutcome::Success(o),
+                            Err(e) => TransactionOutcome::Failure(e),
+                        },
+                        application_events,
+                        application_logs,
+                        system_structure,
+                        execution_trace: if execution_trace_enabled {
+                            Some(execution_trace)
+                        } else {
+                            None
+                        },
+                    }),
+                )
+            }
+            TransactionResultType::Reject(reason) => (
+                costing_parameters,
+                costing_module.fee_reserve.finalize().into(),
+                fee_details,
+                TransactionResult::Reject(RejectResult { reason }),
+            ),
+            TransactionResultType::Abort(reason) => (
+                costing_parameters,
+                costing_module.fee_reserve.finalize().into(),
+                fee_details,
+                TransactionResult::Abort(AbortResult { reason }),
+            ),
+        }
+    }
+
 
     fn on_pin_node(&mut self, node_id: &NodeId) -> Result<(), RuntimeError> {
         SystemModuleMixer::on_pin_node(self, node_id)
