+++ radix-engine/src/track/track.rs
@@ -203,7 +203,7 @@ impl<'s, S: SubstateDatabase, M: DatabaseKeyMapper + 'static> Track<'s, S, M> {
     /// Finalizes changes captured by this substate store.
     ///
     ///  Returns the state changes and dependencies.
-    pub fn finalize(mut self) -> Result<TrackedSubstates, TrackFinalizeError> {
+    pub fn finalize(mut self) -> Result<(TrackedSubstates,  &'s S), TrackFinalizeError> {
         for (node_id, transient_substates) in self.transient_substates.transient_substates {
             for (partition, substate_key) in transient_substates {
                 if let Some(tracked_partition) = self
@@ -224,10 +224,10 @@ impl<'s, S: SubstateDatabase, M: DatabaseKeyMapper + 'static> Track<'s, S, M> {
             }
         }
 
-        Ok(TrackedSubstates {
+        Ok((TrackedSubstates {
             tracked_nodes: self.tracked_nodes,
             deleted_partitions: self.deleted_partitions,
-        })
+        }, self.substate_db))
     }
 
     fn get_tracked_partition(
