+++ radix-engine/src/transaction/transaction_executor.rs
@@ -393,23 +393,6 @@ where
         receipt
     }
 
-    pub fn read_epoch<K: SubstateDatabase>(track: &mut Track<K, SpreadPrefixKeyMapper>) -> Option<Epoch> {
-        // TODO - Instead of doing a check of the exact epoch, we could do a check in range [X, Y]
-        //        Which could allow for better caching of transaction validity over epoch boundaries
-        match track.read_substate(
-            CONSENSUS_MANAGER.as_node_id(),
-            MAIN_BASE_PARTITION,
-            &ConsensusManagerField::State.into(),
-        ) {
-            Some(x) => {
-                let substate: FieldSubstate<ConsensusManagerStateFieldPayload> =
-                    x.as_typed().unwrap();
-                Some(substate.into_payload().into_latest().epoch)
-            }
-            None => None,
-        }
-    }
-
     fn interpret_manifest<T: WrappedSystem<V>>(
         &self,
         mut track: Track<S, SpreadPrefixKeyMapper>,
@@ -462,527 +445,8 @@ where
 
         let system = wrapped_system.to_system();
 
-        #[cfg(not(feature = "alloc"))]
-        if system.modules.enabled_modules.contains(EnabledModules::KERNEL_TRACE) {
-            println!("{:-^120}", "Interpretation Results");
-            println!("{:?}", interpretation_result);
-        }
+        system.on_teardown3(track, executable, interpretation_result)
 
-        let (mut costing_module, runtime_module, execution_trace_module) = system.modules.unpack();
-
-        let costing_parameters = costing_module.fee_reserve.costing_parameters();
-
-        let fee_details = if let Some(cost_breakdown) = &costing_module.cost_breakdown {
-            let cost_breakdown = cost_breakdown.clone();
-            let execution_cost_breakdown = cost_breakdown
-                .execution_cost_breakdown
-                .into_iter()
-                .map(|(k, v)| (k.to_string(), v))
-                .collect();
-            let finalization_cost_breakdown = cost_breakdown
-                .finalization_cost_breakdown
-                .into_iter()
-                .map(|(k, v)| (k.to_string(), v))
-                .collect();
-            Some(TransactionFeeDetails {
-                execution_cost_breakdown,
-                finalization_cost_breakdown,
-            })
-        } else {
-            None
-        };
-
-        let result_type = Self::determine_result_type(
-            interpretation_result,
-            &mut costing_module.fee_reserve,
-        );
-        match result_type {
-            TransactionResultType::Commit(outcome) => {
-                let is_success = outcome.is_ok();
-
-                // Commit/revert
-                if !is_success {
-                    costing_module.fee_reserve.revert_royalty();
-                    track.revert_non_force_write_changes();
-                }
-
-                // Distribute fees
-                let (fee_reserve_finalization, paying_vaults, finalization_events) =
-                    Self::finalize_fees(
-                        &mut track,
-                        costing_module.fee_reserve,
-                        is_success,
-                        executable.costing_parameters().free_credit_in_xrd,
-                    );
-                let fee_destination = FeeDestination {
-                    to_proposer: fee_reserve_finalization.to_proposer_amount(),
-                    to_validator_set: fee_reserve_finalization.to_validator_set_amount(),
-                    to_burn: fee_reserve_finalization.to_burn_amount(),
-                    to_royalty_recipients: fee_reserve_finalization
-                        .royalty_cost_breakdown
-                        .clone(),
-                };
-
-                // Update intent hash status
-                if let Some(next_epoch) = Self::read_epoch(&mut track) {
-                    Self::update_transaction_tracker(
-                        &mut track,
-                        next_epoch,
-                        executable.intent_hash(),
-                        is_success,
-                    );
-                }
-
-                // Finalize events and logs
-                let (mut application_events, application_logs) =
-                    runtime_module.finalize(is_success);
-                application_events.extend(finalization_events);
-
-                // Finalize execution trace
-                let execution_trace =
-                    execution_trace_module.finalize(&paying_vaults, is_success);
-
-                // Finalize track
-                let tracked_substates = {
-                    match track.finalize() {
-                        Ok(result) => result,
-                        Err(TrackFinalizeError::TransientSubstateOwnsNode) => {
-                            panic!("System invariants should prevent transient substate from owning nodes");
-                        }
-                    }
-                };
-
-                // Generate state updates from tracked substates
-                // Note that this process will prune invalid reads
-                let (new_node_ids, state_updates) =
-                    to_state_updates::<SpreadPrefixKeyMapper>(tracked_substates);
-
-                // Summarizes state updates
-                let system_structure = SystemStructure::resolve(
-                    self.substate_db,
-                    &state_updates,
-                    &application_events,
-                );
-                let state_update_summary =
-                    StateUpdateSummary::new(self.substate_db, new_node_ids, &state_updates);
-
-                // Resource reconciliation does not currently work in preview mode
-                if executable.costing_parameters().free_credit_in_xrd.is_zero() {
-                    let system_reader = SystemDatabaseReader::new_with_overlay(
-                        self.substate_db,
-                        &state_updates,
-                    );
-                    reconcile_resource_state_and_events(
-                        &state_update_summary,
-                        &application_events,
-                        system_reader,
-                    );
-                }
-
-                (
-                    costing_parameters,
-                    fee_reserve_finalization.into(),
-                    fee_details,
-                    TransactionResult::Commit(CommitResult {
-                        state_updates,
-                        state_update_summary,
-                        fee_source: FeeSource { paying_vaults },
-                        fee_destination,
-                        outcome: match outcome {
-                            Ok(o) => TransactionOutcome::Success(o),
-                            Err(e) => TransactionOutcome::Failure(e),
-                        },
-                        application_events,
-                        application_logs,
-                        system_structure,
-                        execution_trace: if self.system_init.execution_trace.is_some() {
-                            Some(execution_trace)
-                        } else {
-                            None
-                        },
-                    }),
-                )
-            }
-            TransactionResultType::Reject(reason) => (
-                costing_parameters,
-                costing_module.fee_reserve.finalize().into(),
-                fee_details,
-                TransactionResult::Reject(RejectResult { reason }),
-            ),
-            TransactionResultType::Abort(reason) => (
-                costing_parameters,
-                costing_module.fee_reserve.finalize().into(),
-                fee_details,
-                TransactionResult::Abort(AbortResult { reason }),
-            ),
-        }
-    }
-
-    fn determine_result_type(
-        interpretation_result: Result<Vec<InstructionOutput>, TransactionExecutionError>,
-        fee_reserve: &mut SystemLoanFeeReserve,
-    ) -> TransactionResultType {
-        // A `SuccessButFeeLoanNotRepaid` error is issued if a transaction finishes before
-        // the SYSTEM_LOAN_AMOUNT is reached (which trigger a repay event) and even though
-        // enough fee has been locked.
-        //
-        // Do another `repay` try during finalization to remedy it.
-        let final_repay_result = fee_reserve.repay_all();
-
-        match interpretation_result {
-            Ok(output) => match final_repay_result {
-                Ok(_) => TransactionResultType::Commit(Ok(output)), // success and system loan repaid fully
-                Err(e) => {
-                    if let Some(abort_reason) = e.abortion() {
-                        TransactionResultType::Abort(abort_reason.clone())
-                    } else {
-                        TransactionResultType::Reject(RejectionReason::SuccessButFeeLoanNotRepaid)
-                    }
-                }
-            },
-            Err(e) => match e {
-                TransactionExecutionError::BootloadingError(e) => {
-                    TransactionResultType::Reject(RejectionReason::BootloadingError(e))
-                }
-                TransactionExecutionError::RuntimeError(e) => {
-                    if let Some(abort_reason) = e.abortion() {
-                        TransactionResultType::Abort(abort_reason.clone())
-                    } else {
-                        if fee_reserve.fully_repaid() {
-                            TransactionResultType::Commit(Err(e))
-                        } else {
-                            TransactionResultType::Reject(
-                                RejectionReason::ErrorBeforeLoanAndDeferredCostsRepaid(e),
-                            )
-                        }
-                    }
-                }
-            },
-        }
-    }
-
-    fn finalize_fees(
-        track: &mut Track<S, SpreadPrefixKeyMapper>,
-        fee_reserve: SystemLoanFeeReserve,
-        is_success: bool,
-        free_credit: Decimal,
-    ) -> (
-        FeeReserveFinalizationSummary,
-        IndexMap<NodeId, Decimal>,
-        Vec<(EventTypeIdentifier, Vec<u8>)>,
-    ) {
-        let mut events = Vec::<(EventTypeIdentifier, Vec<u8>)>::new();
-
-        // Distribute royalty
-        for (recipient, amount) in fee_reserve.royalty_cost_breakdown().clone() {
-            let node_id = recipient.vault_id();
-            let substate_key = FungibleVaultField::Balance.into();
-            let mut vault_balance = track
-                .read_substate(&node_id, MAIN_BASE_PARTITION, &substate_key)
-                .unwrap()
-                .as_typed::<FungibleVaultBalanceFieldSubstate>()
-                .unwrap()
-                .into_payload()
-                .into_latest();
-            vault_balance.put(LiquidFungibleResource::new(amount));
-            let updated_substate_content =
-                FungibleVaultBalanceFieldPayload::from_content_source(vault_balance)
-                    .into_unlocked_substate();
-            track
-                .set_substate(
-                    node_id,
-                    MAIN_BASE_PARTITION,
-                    substate_key,
-                    IndexedScryptoValue::from_typed(&updated_substate_content),
-                    &mut |_| -> Result<(), ()> { Ok(()) },
-                )
-                .unwrap();
-            events.push((
-                EventTypeIdentifier(
-                    Emitter::Method(node_id, ModuleId::Main),
-                    DepositEvent::EVENT_NAME.to_string(),
-                ),
-                scrypto_encode(&DepositEvent { amount }).unwrap(),
-            ));
-        }
-
-        // Take fee payments
-        let fee_reserve_finalization = fee_reserve.finalize();
-        let mut fee_payments: IndexMap<NodeId, Decimal> = index_map_new();
-        let mut required = fee_reserve_finalization.total_cost();
-        let mut collected_fees = LiquidFungibleResource::new(Decimal::ZERO);
-        for (vault_id, mut locked, contingent) in
-            fee_reserve_finalization.locked_fees.iter().cloned().rev()
-        {
-            let amount = if contingent {
-                if is_success {
-                    Decimal::min(locked.amount(), required)
-                } else {
-                    Decimal::zero()
-                }
-            } else {
-                Decimal::min(locked.amount(), required)
-            };
-
-            // NOTE: Decimal arithmetic operation safe unwrap.
-            // No chance to overflow considering current costing parameters
-
-            // Take fees
-            collected_fees.put(locked.take_by_amount(amount).unwrap());
-            required = required.checked_sub(amount).unwrap();
-
-            // Refund overpayment
-            let mut vault_balance = track
-                .read_substate(
-                    &vault_id,
-                    MAIN_BASE_PARTITION,
-                    &FungibleVaultField::Balance.into(),
-                )
-                .unwrap()
-                .as_typed::<FungibleVaultBalanceFieldSubstate>()
-                .unwrap()
-                .into_payload()
-                .into_latest();
-            vault_balance.put(locked);
-            let updated_substate_content =
-                FungibleVaultBalanceFieldPayload::from_content_source(vault_balance)
-                    .into_unlocked_substate();
-            track
-                .set_substate(
-                    vault_id,
-                    MAIN_BASE_PARTITION,
-                    FungibleVaultField::Balance.into(),
-                    IndexedScryptoValue::from_typed(&updated_substate_content),
-                    &mut |_| -> Result<(), ()> { Ok(()) },
-                )
-                .unwrap();
-
-            // Record final payments
-            let entry = fee_payments.entry(vault_id).or_default();
-            *entry = entry.checked_add(amount).unwrap();
-
-            events.push((
-                EventTypeIdentifier(
-                    Emitter::Method(vault_id, ModuleId::Main),
-                    PayFeeEvent::EVENT_NAME.to_string(),
-                ),
-                scrypto_encode(&PayFeeEvent { amount }).unwrap(),
-            ));
-        }
-        // Free credit is locked first and thus used last
-        if free_credit.is_positive() {
-            let amount = Decimal::min(free_credit, required);
-            collected_fees.put(LiquidFungibleResource::new(amount));
-            required = required.checked_sub(amount).unwrap();
-        }
-
-        let to_proposer = fee_reserve_finalization.to_proposer_amount();
-        let to_validator_set = fee_reserve_finalization.to_validator_set_amount();
-        let to_burn = fee_reserve_finalization.to_burn_amount();
-
-        // Sanity checks
-        assert!(
-            fee_reserve_finalization.total_bad_debt_in_xrd == Decimal::ZERO,
-            "Bad debt is non-zero: {}",
-            fee_reserve_finalization.total_bad_debt_in_xrd
-        );
-        assert!(
-            required == Decimal::ZERO,
-            "Locked fee does not cover transaction cost: {} required",
-            required
-        );
-        let remaining_collected_fees = collected_fees.amount().checked_sub(fee_reserve_finalization.total_royalty_cost_in_xrd /* royalty already distributed */).unwrap();
-        let to_distribute = to_proposer
-            .checked_add(to_validator_set)
-            .unwrap()
-            .checked_add(to_burn)
-            .unwrap();
-        assert!(
-            remaining_collected_fees  == to_distribute,
-            "Remaining collected fee isn't equal to amount to distribute (proposer/validator set/burn): {} != {}",
-            remaining_collected_fees,
-            to_distribute,
-        );
-
-        if !to_proposer.is_zero() || !to_validator_set.is_zero() {
-            // Fetch current leader
-            // TODO: maybe we should move current leader into validator rewards?
-            let substate: FieldSubstate<ConsensusManagerStateFieldPayload> = track
-                .read_substate(
-                    CONSENSUS_MANAGER.as_node_id(),
-                    MAIN_BASE_PARTITION,
-                    &ConsensusManagerField::State.into(),
-                )
-                .unwrap()
-                .as_typed()
-                .unwrap();
-            let current_leader = substate.into_payload().into_latest().current_leader;
-
-            // Update validator rewards
-            let substate: FieldSubstate<ConsensusManagerValidatorRewardsFieldPayload> = track
-                .read_substate(
-                    CONSENSUS_MANAGER.as_node_id(),
-                    MAIN_BASE_PARTITION,
-                    &ConsensusManagerField::ValidatorRewards.into(),
-                )
-                .unwrap()
-                .as_typed()
-                .unwrap();
-
-            let mut rewards = substate.into_payload().into_latest();
-
-            if let Some(current_leader) = current_leader {
-                let entry = rewards.proposer_rewards.entry(current_leader).or_default();
-                *entry = entry.checked_add(to_proposer).unwrap()
-            } else {
-                // If there is no current leader, the rewards go to the pool
-            };
-            let vault_node_id = rewards.rewards_vault.0 .0;
-
-            track
-                .set_substate(
-                    CONSENSUS_MANAGER.into_node_id(),
-                    MAIN_BASE_PARTITION,
-                    ConsensusManagerField::ValidatorRewards.into(),
-                    IndexedScryptoValue::from_typed(&FieldSubstate::new_unlocked_field(
-                        ConsensusManagerValidatorRewardsFieldPayload::from_content_source(rewards),
-                    )),
-                    &mut |_| -> Result<(), ()> { Ok(()) },
-                )
-                .unwrap();
-
-            // Put validator rewards into the vault
-            let total_amount = to_proposer.checked_add(to_validator_set).unwrap();
-            let mut vault_balance = track
-                .read_substate(
-                    &vault_node_id,
-                    MAIN_BASE_PARTITION,
-                    &FungibleVaultField::Balance.into(),
-                )
-                .unwrap()
-                .as_typed::<FungibleVaultBalanceFieldSubstate>()
-                .unwrap()
-                .into_payload()
-                .into_latest();
-            vault_balance.put(collected_fees.take_by_amount(total_amount).unwrap());
-            let updated_substate_content =
-                FungibleVaultBalanceFieldPayload::from_content_source(vault_balance)
-                    .into_unlocked_substate();
-            track
-                .set_substate(
-                    vault_node_id,
-                    MAIN_BASE_PARTITION,
-                    FungibleVaultField::Balance.into(),
-                    IndexedScryptoValue::from_typed(&updated_substate_content),
-                    &mut |_| -> Result<(), ()> { Ok(()) },
-                )
-                .unwrap();
-
-            events.push((
-                EventTypeIdentifier(
-                    Emitter::Method(vault_node_id, ModuleId::Main),
-                    DepositEvent::EVENT_NAME.to_string(),
-                ),
-                scrypto_encode(&DepositEvent {
-                    amount: total_amount,
-                })
-                .unwrap(),
-            ));
-        }
-
-        if to_burn.is_positive() {
-            events.push((
-                EventTypeIdentifier(
-                    Emitter::Method(XRD.into_node_id(), ModuleId::Main),
-                    "BurnFungibleResourceEvent".to_string(),
-                ),
-                scrypto_encode(&BurnFungibleResourceEvent { amount: to_burn }).unwrap(),
-            ));
-        }
-
-        (fee_reserve_finalization, fee_payments, events)
-    }
-
-    fn update_transaction_tracker(
-        track: &mut Track<S, SpreadPrefixKeyMapper>,
-        next_epoch: Epoch,
-        intent_hash: &TransactionIntentHash,
-        is_success: bool,
-    ) {
-        // Read the intent hash store
-        let transaction_tracker = track
-            .read_substate(
-                TRANSACTION_TRACKER.as_node_id(),
-                MAIN_BASE_PARTITION,
-                &TransactionTrackerField::TransactionTracker.into(),
-            )
-            .unwrap()
-            .as_typed::<FieldSubstate<TransactionTrackerSubstate>>()
-            .unwrap()
-            .into_payload();
-
-        let mut transaction_tracker = transaction_tracker.into_v1();
-
-        // Update the status of the intent hash
-        if let TransactionIntentHash::ToCheck {
-            expiry_epoch,
-            intent_hash,
-        } = intent_hash
-        {
-            if let Some(partition_number) =
-                transaction_tracker.partition_for_expiry_epoch(*expiry_epoch)
-            {
-                track
-                    .set_substate(
-                        TRANSACTION_TRACKER.into_node_id(),
-                        PartitionNumber(partition_number),
-                        SubstateKey::Map(scrypto_encode(intent_hash).unwrap()),
-                        IndexedScryptoValue::from_typed(&KeyValueEntrySubstate::V1(
-                            KeyValueEntrySubstateV1 {
-                                value: Some(if is_success {
-                                    TransactionStatus::V1(TransactionStatusV1::CommittedSuccess)
-                                } else {
-                                    TransactionStatus::V1(TransactionStatusV1::CommittedFailure)
-                                }),
-                                // TODO: maybe make it immutable, but how does this affect partition deletion?
-                                lock_status: LockStatus::Unlocked,
-                            },
-                        )),
-                        &mut |_| -> Result<(), ()> { Ok(()) },
-                    )
-                    .unwrap();
-            } else {
-                panic!("No partition for an expiry epoch")
-            }
-        }
-
-        // Check if all intent hashes in the first epoch have expired, based on the `next_epoch`.
-        //
-        // In this particular implementation, because the transaction tracker coverage is greater than
-        // the max epoch range in transaction header, we must check epoch range first to
-        // ensure we don't store intent hash too far into the future.
-        //
-        // Also, we need to make sure epoch doesn't jump by a large distance.
-        if next_epoch.number()
-            >= transaction_tracker.start_epoch + transaction_tracker.epochs_per_partition
-        {
-            let discarded_partition = transaction_tracker.advance();
-            track.delete_partition(
-                TRANSACTION_TRACKER.as_node_id(),
-                PartitionNumber(discarded_partition),
-            );
-        }
-        track
-            .set_substate(
-                TRANSACTION_TRACKER.into_node_id(),
-                MAIN_BASE_PARTITION,
-                TransactionTrackerField::TransactionTracker.into(),
-                IndexedScryptoValue::from_typed(&FieldSubstate::new_unlocked_field(
-                    TransactionTrackerSubstate::V1(transaction_tracker),
-                )),
-                &mut |_| -> Result<(), ()> { Ok(()) },
-            )
-            .unwrap();
     }
 
     #[cfg(not(feature = "alloc"))]
@@ -1161,7 +625,7 @@ pub fn execute_and_commit_transaction<
     receipt
 }
 
-enum TransactionResultType {
+pub enum TransactionResultType {
     Commit(Result<Vec<InstructionOutput>, RuntimeError>),
     Reject(RejectionReason),
     Abort(AbortReason),
